requirements.txt
apturl==0.5.2
awscli==1.11.13
beautifulsoup4==4.4.1
blinker==1.3
botocore==1.4.70
Brlapi==0.6.4
chardet==2.3.0
checkbox-support==0.22
colorama==0.3.7
command-not-found==0.3
cryptography==1.2.3
defer==1.0.6
dj-database-url==0.4.2
Django==1.10.6
djangorestframework==3.6.2
docutils==0.12
feedparser==5.1.3
guacamole==0.9.2
gunicorn==19.7.1
html5lib==0.999
httplib2==0.9.1
idna==2.0
Jinja2==2.8
jmespath==0.9.0
language-selector==0.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
MarkupSafe==0.23
oauthlib==1.0.3
onboard==1.2.0
padme==1.1.1
pexpect==4.0.1
Pillow==3.1.2
plainbox==0.25
psycopg2==2.7.1
ptyprocess==0.5
pyasn1==0.1.9
pycups==1.9.73
pycurl==7.43.0
Pygments==2.1
pygobject==3.20.0
PyJWT==1.3.0
pyparsing==2.0.3
python-apt==1.1.0b1
python-dateutil==2.4.2
python-debian==0.1.27
python-systemd==231
pyxdg==0.25
reportlab==3.3.0
requests==2.9.1
roman==2.0.0
rsa==3.2.3
s3transfer==0.1.9
sessioninstaller==0.0.0
six==1.10.0
system-service==0.3
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.13.1
usb-creator==0.3.0
whitenoise==3.3.0
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
